Calculate upgrade1 score, upgrade2 score and uptime count of validators

+++ util/uptime/main.go
@@ -44,5 +44,5 @@ func main() {
 
 	handler := src.New(session)
 
-	handler.CalculateUptime(startBlock, endBlock)
+	handler.CalculateUptime(int64(startBlock), int64(endBlock))
 }
